pyproject.toml
[project]
name = "cinema"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==4.2",
    "django-filter>=25.1",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt[crypto]>=5.5.1",
    "drf-spectacular>=0.28.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "requests>=2.32.5",
]
